Fixes. LOTS of fixes
